refactor: use table for config.pipe
This necessitated a fairly significant refactor of the code that checks + applies
user-supplied options to handle sub-tables in a robust way. The end result
is powerful enough to handle arbitrary levels of nesting, and doesn't
require a user to specify every option in a subtable if they don't like the
default.

To make this as simple as possible, all options (including hook elements)
must now be non-nil.

fix: correctly apply rproj settings for pipe version